Add a `proj` subcommand to `subl`
Sublime's "project" tool lets you close a window and reopen it later
with all tabs intact. It probably has other features but that's what I
want right now.

Normally to start a project for the first time you have to
1. have a window open already
2. go to the menu bar; (there's no keyboard shortcut)

Having a tiny skeleton of a project is sufficient to make Sublime do the
needful. After that, you just need to be sure to open a window with
`--project`. This streamlines that task.