Fix subl_wait configuration
The most important bit of this change is using an explicit path when
setting EDITOR rather than relying on `which`. Since I have `bin` near
the front of my path, `which subl_wait.fish` returns
`bin/subl_wait.fish`, which isn't useful once I type cd.

`-x subl_wait.fish` is also an ineffective test; it'll always be there
because it's in source control. The important question is whether `subl`
is a thing--i.e. whether sublime is available.